requirements.txt
numpy==1.24.0
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
Pillow==10.0.0
scikit-image==0.21.0
plotly==5.17.0
